+++ ai-logic/firebase-ai/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- [changed] Remove deprecated classes and apis. (#8563)
+- [changed] Restrict constructors on output/status types to internal. (#8563)
+- [changed] Remove vertex aI and use agent platform instead. (#8563)
+
 # 17.17.0
 
 - [changed] Added `turnComplete` support in the `LiveSessionFutures` Java API. (#8550)